整理OAuth 使用middleware
